Implement rhino.evaluate() to evaluate scripts on an existing scope
This may be useful for emulating "traditional" JavaScript environments
such as browsers, where all scripts are evaluated against the same global
scope.

The first argument is a module name, the optional second one a JavaScript
object to be used as evaluation scope. If no second argument is given, a
simple top level scope is created. The function returns the scope object.

git-svn-id: https://dev.helma.org/svn/helma-ng/trunk@9151 688a9155-6ab5-4160-a077-9df41f55a9e9