Implement module exports as defined in ServerJS Securable Modules proposal
https://wiki.mozilla.org/ServerJS/Modules/SecurableModules

This means that require() no longer returns the module scope itself but the exports
object, containing only the module's exported properties. The global export() is now
implemented on top of this and continues to work as before. As a consequence, export
definitions no longer only apply to include() but also require() and import(), and an
ommitted export causes the exported property not to be available at all.

This means explicit exports are now necessary where they weren't before: in the app's
config module, in webapp modules defining action functions, and in test case modules.

As a little reward in return, actions no longer need the _action suffix as only
functions that are explicitly exported are invokable as actions.

git-svn-id: https://dev.helma.org/svn/helma-ng/trunk@9537 688a9155-6ab5-4160-a077-9df41f55a9e9