Fix `ringo-admin create -a` bug on win32
We currently do the following when creating the appengine directory
structure:

    makeDirectory("static")
    move(..., "static")

This fails on Win32. The underlying issue is that rename/move semantics
are platform-dependent: POSIX is happy to overwrite an existing target
under certain circumstances; Win32 fails always if the target already
exists.

Fortunately, in our case the issue is easily fixed. As we know that the
source for the `move` always exists, we also know that the target will
always exist, therefore we can just drop the `makeDirectory`.