Implement deep module loading from packages.
If a module id consists of multiple terms and the first term resolves to
a package directory, the remaining module id will be resolved against the
"lib" directory of the package (configurable in package.json via
"directories": {"lib": "..."}).

If a module id consists of a single term and resolves to a package directory,
the main module of the package is returned as usual.