fixed a tricky bug in scheduler test module:
before: when executing `ringo test all.js` this module would lead to a mass of test failures. when executing `ringo all.js` everything is fine.

reason: the `scheduler_test` module instantiates itself in a new worker - it requires the `assert` module, which in turn requires the test module. when the test module is required, `system.args` contains `[test,all.js]`, which basically means the whole test suite is executed again. this leads to unpredictable behaviour, where suddently methods like `jsDump` and `getType` (exported by test module) are undefined.

the solution i've chosen is to move the worker code into a separate module, and use this module to instantiate the workers used in the test. additional benefit: the scheduler test can be called from commandline again (even with fixing the missing require of system module this lead to a lock).