assert.throws() no longer accepts String as expectedError
I just wanted to add the optional `message` argument assert.throws but the spec
is broken in that regard; e.g., with `throws(fooFunction, 'foo')` it is unclear
whether `foo` is the expected error to be thrown or the optional message arg.

Ringo now only allows `Function` or `RegExp` as `expectedError` types.
This is in line with qunit and nodejs and it allows us to finally have that 
message argument.