fix1:globals.js setInterval bug caused by new thread not able to retrieve currentWorker from ThreadLocal fix2:swing.js example need to remove require(scheduler) statement as setInterval has been moved to globals.js
in globals.js line 124
 ```javascript
        var worker = engine.getCurrentWorker();
```
this line will get a null object back if the current thread is created without a worker attached to it.
example is the swing.js example when a mouse click on it, that thread is created in java , therefore no worker associated with it. when a listener is running in that thread, you will not be able to retrieve a worker from ThreadLocal.
Therefore code fix is to test if current thread has a worker , if not, create one to run the code