Drops the Permission class
The CommonJS spec is quite vague about the Permission class. Since only Posix is supported and all common use cases for this include octal representation, we should drop this class. Using umask isn't possible in a Java-only fs module, so one of the justifiable features of the Permission class also disappears.