Adapts main launcher for Java 9+ class loading
Fixes #417. Starting with Java 9 the AppClassLoader no longer is an instance of the URLClassLoader. Since only the URL to the main class is needed to resolve a valid RINGO_HOME, we can switch to the standard getResource(), which will walk up the class loader hierarchy. The old findResource() approach only used the current class loader to find a resource.

Improves the initial fix in b21d8908dfb98a5208b0e033b817838a06dde111