use the main worker in ringo shell instead of a separate new worker
reason: when loading a module in interactive mode (i.e. `ringo -i mymodule.js`) the module would be evaluated
using the main worker, but any calls of the module methods would be executed in the shell worker, which is
a different one. this makes using eg. `setTimeout` in module methods impossible because they fail with
`Current thread worker differs from scope worker` exception